requirements.txt
google-adk==1.0.0
mcp==1.9.1
google-generativeai==0.8.5
python-dotenv==1.1.0
requests==2.32.3
deprecated==1.2.14
